requirements.txt
# aiohttp==1.0.5
# discord.py==0.16.12
aiohttp==3.6.0
discord.py==1.4.0
async-timeout==3.0.1
cffi==1.12.2
chardet==3.0.4
multidict==4.5.2
pycparser==2.19
PyNaCl==1.1.2
six==1.12.0
websockets==3.4
